--- src/main/java/Command.java
@@ -1,6 +1,0 @@
-public abstract class Command {
-
-    abstract void execute(TaskList tasks, Ui ui, Storage storage);
-    abstract boolean isExit();
-
-}
