--- src/main/java/ListCommand.java
@@ -1,12 +1,0 @@
-public class ListCommand extends Command {
-
-    @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.printTaskList();
-    }
-
-    @Override
-    boolean isExit() {
-        return false;
-    }
-}
