+++ src/main/java/duke/commands/AddCommand.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class AddCommand extends Command {
 
     private char taskType;
@@ -9,7 +15,7 @@ public class AddCommand extends Command {
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (taskType == 'T') {
             tasks.addTodo(userInput);
         } else if (taskType == 'D') {
@@ -20,7 +26,7 @@ public class AddCommand extends Command {
     }
 
     @Override
-    boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 }
