+++ src/main/java/duke/commands/ListCommand.java
@@ -1,0 +1,18 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class ListCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        tasks.printTaskList();
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
