+++ src/main/java/duke/events/Deadlines.java
@@ -1,3 +1,5 @@
+package duke.events;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
