+++ src/main/java/duke/events/Events.java
@@ -1,3 +1,5 @@
+package duke.events;
+
 public class Events extends Task {
 
     private String taskName;
